Refactor:[파일 이동] utils 폴더 내 services 폴더로 api 호출 함수 삽입
