+++ src/app/sign-up/page.tsx
@@ -7,7 +7,7 @@ import { useRouter } from "next/navigation";
 import React from "react";
 import { useForm } from "react-hook-form";
 import { ICommonResponse } from "@/typescript/common/response.interface";
-import { ImageUpload } from "@/utils/upload";
+import { ImageUpload } from "@/utils/services/upload";
 
 /**
  * 회원가입 페이지
